+++ src/once.ts
@@ -3,7 +3,7 @@
  * If _after_ function is provided, then
  * it will be called after _fn_ has been executed.
  */
-export const once = <F extends (...args: ReadonlyArray<unknown>) => unknown>(
+export const once = <F extends (...args: ReadonlyArray<any>) => any>(
   fn: F,
   after?: F
 ): F => {